requirements.txt
fastapi
uvicorn
cassandra-driver
python-dotenv
email-validator
argon2-cffi
jinja2
python-multipart
python-jose[cryptography]
algoliasearch
chardet